+++ wpipe_steps/huggingface/__init__.py
@@ -1,0 +1,1 @@
+from .text.classification import HFTextClassificationStep